% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charcoal-package.R
\docType{package}
\name{charcoal}
\alias{charcoal}
\title{charcoal: An R package implementing the charcoal algorithms for localizing
changepoints in high-dimensional linear regression (via complementary
sketching)}
\description{
The charcoal package collects the implementation for the novel 
methodology #' 'charcoal' for changepoint localization in high-dimensional 
linear regression, introduced in Gao and Wang (2022).
}
\references{
Gao, F. and Wang, T. (2022) Sparse change detection in
high-dimensional linear regression. arXiv preprint, arXiv:2208.06326.
}
