% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{cpreg}
\alias{cpreg}
\title{Main function implementing the charcoal algorithm in Gao and Wang (2022)}
\usage{
cpreg(X, Y, lambda = NA, sigma = NULL, aggregate_method = 1)
}
\arguments{
\item{X}{design matrix of the linear regression model}

\item{sigma}{noise standard deviation of the regression models; if unknown, set to NULL and will be estimated.}

\item{aggregate_method}{specifies which variant of the algorithms to be applied. Set it to 1 for Algorithm 1 (default), and to 2 for Algorithm 2.}

\item{y}{response vector of the linear regression model}

\item{sparse}{whether to test a sparse difference in the regression coefficients of the two models, default is TRUE}
}
\value{
a list containing the test statistics and changepoint estimate.
}
\description{
From the model y_t = x_t beta_1 + eps_t for 1<=t<=z and y_t = x_t beta_2 + eps_t for z+1<=t<=n, where x_t are p-dimensional design vectors for 1<=t<=n, we estimate the changepoint z where the change of regression coefficient takes place. The flag sparse is set to be TRUE if beta1 - beta2 has at most sqrt(p) so that appropriate testing statistics are formed. The test is performed using the charcoal algorithm from Gao and Wang (2022).
}
\examples{
# problem parameters
n <- 240 # total sample size
z <- 150  # changepoint location
p <- 160 # dimension of covariates
k <- 10 # sparsity of difference of the two regression coefficients
rho <- 1 # difference in l_2 norm of the two regression coefficients
# generate design matrices
X1 <- matrix(rnorm(z * p), z, p)
X2 <- matrix(rnorm((n-z) * p), (n-z), p)
# generate regression coefficients
beta1 <- rnorm(p)
theta <- c(rnorm(k), rep(0, p-k))
theta <- theta / sqrt(sum(theta^2)) * rho
beta2 <- beta1 + theta
# generate response vectors
y1 <- X1 \%*\% beta1 + rnorm(n1)
y2 <- X2 \%*\% beta2 + rnorm(n2)
# combine the samples before and after changepoint to form the response and covariates
X <- rbind(X1,X2)
Y <- rbind(y1,y2)
# test for difference in beta1 and beta2
result <- cpreg(X,Y)
# or  result <- cpreg(X,Y,lambda=NA,sigma=NULL,aggregate_method=2)
}
\references{
Gao, F. and Wang, T. (2022) Sparse change detection in high-dimensional regression. Work-in-progress.
}
