% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{dicker_noise_sd}
\alias{dicker_noise_sd}
\title{Computing the noise variance in a high-dimensional linear model}
\usage{
dicker_noise_sd(W, z)
}
\arguments{
\item{W}{design matrix for high-dimensional linear model}

\item{z}{response in a high-dimensional linear model}
}
\value{
a nonnegative scalar of estimated noise standard deviation
}
\description{
Assume data are generated from the model z = W b + e, where e
has independent components N(0, s^2), we estimate s^2 by using the method
in Dicker (2014).  If the method fails, a fallback with nosie_sd will be
used.
}
\references{
Dicker, L. H. (2014). Variance estimation in high-dimensional
linear models. Biometrika, 101(2), 269-284.
}
