% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{getNOTIntervals}
\alias{getNOTIntervals}
\title{Internal function to generate random intervals for `not_cpreg`, the main
multiple changepoint estimation function in linear regressions as in
Algorithm 4}
\usage{
getNOTIntervals(n, p, no_intervals, delta = 0.1)
}
\arguments{
\item{n}{the sample size in question}

\item{p}{the regression coefficient dimension in question}

\item{no_intervals}{the desired number of intervals}

\item{delta}{only intervals with length longer than (1+delta) * p will be
kept}
}
\description{
Internal function to generate random intervals for `not_cpreg`, the main
multiple changepoint estimation function in linear regressions as in
Algorithm 4
}
