% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{matrix.power}
\alias{matrix.power}
\title{Compute matrix power of a symmetric matrix}
\usage{
matrix.power(A, power, pseudoinverse = TRUE)
}
\arguments{
\item{A}{a square matrix}

\item{power}{power exponent}

\item{pseudoinverse}{whether to use pseudoinverse if power is negative}
}
\description{
Compute matrix power of a symmetric matrix
}
