% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{noise_sd}
\alias{noise_sd}
\title{Computing the noise variance in a high-dimensional linear model}
\usage{
noise_sd(W, z)
}
\arguments{
\item{W}{design matrix for high-dimensional linear model}

\item{z}{response in a high-dimensional linear model}
}
\value{
a nonnegative scalar of estimated noise standard deviation
}
\description{
Assume data are generated from the model z = W beta + e, where e has independent components N(0, sigma^2), we estimate sigma^2 by fitting a cross-validated Lasso estimator of beta and then estimate sigma from the residual sum of squares.
}
