% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{orthogonalProjection}
\alias{orthogonalProjection}
\title{Computing the orthogonal matrix spanning the orthogonal complement of the
range of X}
\usage{
orthogonalProjection(X)
}
\arguments{
\item{X}{an n x p matrix with n > p}
}
\value{
matrix A of size n x (n-p) with orthogonal columns
}
\description{
Generate the orthogonal complement of the range of matrix X
}
