% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancillary.R
\name{printPercentage}
\alias{printPercentage}
\title{Print percentage}
\usage{
printPercentage(ind, tot)
}
\arguments{
\item{ind}{a vector of for loop iterator}

\item{tot}{a vector of for loop length}
}
\value{
on screen output of percentage
}
\description{
From github.com/wangtengyao/putils
}
