% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{simulateMultiCP}
\alias{simulateMultiCP}
\title{Generate a sequential linear regression sample with a single changepoint or
multiple changepoints}
\usage{
simulateMultiCP(
  n,
  tau,
  p,
  k,
  rho,
  design = "gaussian",
  error = "gaussian",
  sigma = 1,
  seed = NULL
)
}
\arguments{
\item{n}{the total number of data points}

\item{tau}{a number (in case of single changepoint) or vector containing all
changepoints in fraction. If tau = c(0.1,0.2), the changes takes place at
0.1n and 0.2n.  If the lengths of `tau`, `k` and `rho` are all 1, it
generate a linear regression sample with a single changepoint.}

\item{p}{the dimension of regression coefficients}

\item{k}{a vector containing the sparsity of each vector of change, if k is
of length 1, then the same sparsity is applied to all change vectors. E.g.,
k = c(3,5), the first change vector is 3-sparse and the second 5-sparse.}

\item{rho}{signal strength at each changepoint, i.e. the l2 norm of each
change vector, if rho is of length 1, then the signal strength is applied to
all changes.}

\item{design}{a string specifying the distribution of the random design,
implemented options include "gaussian" for GOE design;"t4" for t distribution
 with degree of freedom 4 (4 can be replaced by any desired df, and if
 unspecified, df of 4 is applied); "correlated" or "ar" for each row
 generated by N_p(0,Sigma), where Sigma has (i,j) entry of 0.7^(i-j);
 "anova" for one-way anova design; and "rademacher" for all entries of X
 being independent fair coin tosses between {-1,1}.}

\item{error}{a string specifying the distribution of observational errors
in linear regression, the implemented options include "gaussian" for
standard normal errors;"t4" for t distribution with degree of freedom 4
(4 can be replaced by any desired df, and if unspecified, df of 4 is
applied); "exp" the centered exponential distribution; "laplace" for
standardized Laplace distribution with zero mean and std 1; and "rademacher"
for all errors of being independent fair coin tosses between {-1,1}.}

\item{sigma}{the regression noise level, uniform among all time points}

\item{seed}{the random seed for the ensuing random interval generation for
the narrowest-over-threshold. The default is NULL, where no seed is set.}
}
\value{
a list containing (X, Y, beta, tau). X as the design matrix and Y as
the response vector, beta as the regression coefficient matrix with each
column corresponding to each time point, tau as the fraction of n where
true changes take place
}
\description{
Generate a sequential linear regression sample with multiple
changepoints, with options for different distributions of both random design
and observational error
}
