% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{vector.hard.thresh}
\alias{vector.hard.thresh}
\title{Hard thresholding a vector}
\usage{
vector.hard.thresh(x, lambda)
}
\arguments{
\item{x}{a vector of real numbers}

\item{lambda}{hard thresholding value}
}
\value{
a vector of the same length
}
\description{
entries of `x` that are below lambda are set to 0.
From `github.com/wangtengyao/putils`.
}
