% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charcoal-package.R
\docType{package}
\name{charcoal}
\alias{charcoal}
\title{charcoal: A package for changepoint localization in high-dimensional linear 
regression}
\description{
The charcoal package collects the implementation for the novel methodology
'charcoal' for changepoint localization in high-dimensional linear
regression, introduced in Gao and Wang (2022).
}
