% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{cpreg}
\alias{cpreg}
\title{Main function implementing the charcoal algorithms for single changepoint
in Gao and Wang (2022), including Algorithms 1, 2 and 3}
\usage{
cpreg(
  X,
  Y,
  lambda = NA,
  sigma = NULL,
  burnIn = 0,
  aggregate_method = "proj",
  verbose = FALSE
)
}
\arguments{
\item{X}{design matrix of the linear regression model}

\item{Y}{response vector of the linear regression model}

\item{lambda}{threshold for Algorithms 1 and 2, set it to NULL and the recommended value is applied}

\item{sigma}{noise standard deviation of the regression models; if unknown, set to NULL.}

\item{burnIn}{specifies the fraction at both ends of the interval to be discarded as possible changes, to handle common boundary effects}

\item{aggregate_method}{specifies which variant of the algorithms to be applied. Set it to 'compsket' for Algorithm 1, to 'proj' for Algorithm 2, and to 'lasso_bic' for Algorithm 3.}

\item{verbose}{whether to output intermidate results/progress bar in the console}
}
\value{
a list containing the test statistics and changepoint estimate.
}
\description{
From the model y_t = x_t beta_1 + eps_t for 1<=t<=z and y_t = x_t beta_2 + eps_t for z+1<=t<=n, where x_t are p-dimensional design vectors for 1<=t<=n, we estimate the changepoint z where the change of regression coefficient takes place. The localization is performed using the charcoal algorithms from Gao and Wang (2022).
}
\examples{
# problem parameters
n <- 240 # total sample size
z <- 150  # changepoint location
p <- 160 # dimension of covariates
k <- 10 # sparsity of difference of the two regression coefficients
rho <- 1 # difference in l_2 norm of the two regression coefficients
# generate design matrices
X1 <- matrix(rnorm(z * p), z, p)
X2 <- matrix(rnorm((n-z) * p), (n-z), p)
# generate regression coefficients
beta1 <- rnorm(p)
theta <- c(rnorm(k), rep(0, p-k))
theta <- theta / sqrt(sum(theta^2)) * rho
beta2 <- beta1 + theta
# generate response vectors
y1 <- X1 \%*\% beta1 + rnorm(z)
y2 <- X2 \%*\% beta2 + rnorm(n-z)
# combine the samples before and after changepoint to form the
# response and covariates
X <- rbind(X1,X2)
Y <- rbind(y1,y2)
# look for changepoints
result <- cpreg(X,Y)
# or  result <- cpreg(X,Y,burnIn=0.05,aggregate_method='lasso_bic')
# Alternatively, X and Y can be generated by
ret <- simulateMultiCP(n,z/n,p,k,rho,sigma=1); X <- ret$X; Y <- ret$Y
}
\references{
Gao, F. and Wang, T. (2022) Sparse change detection in high-dimensional linear regression.
}
